mk: add libraries folder as search directory for headers
The next commits will make all include directives of libraries headers be:

 - relative to the libraries directory when the header file is outside the
   source's folder

 - relative to the source's folder when the header file also belongs to that
   directory (directly or indirectly)

One of the reasons for that change is the reduction of search paths (which *may*
make lookups faster in systems like MS Windows [can't check that though, since
I'm a Linux user]).

After the commits that make those changes on include directives, the use of
each library folder in make.inc as search path will be dropped.