AP_Compass: store mag field in the correct unit
This patch additionally fix a problem related to the offsets: now that we're
sending mag field in the correct unit throught mavlink, the offsets received
from the GCS lice calibration are in the same unit. Thus, we need to apply the
unit conversion before applying the offset and other corrections in
AP_Compass_Backend::correct_field().