waf: mavgen: use update_outputs
That makes the output nodes receive build signatures. Since the real output
list (i.e. the header files) is somewhat dynamic because of the messages
description in the XML file, we need to add a post_run method to replace the
outputs variable containing the output dir with the header files.